requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
cattrs==23.2.3
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-allauth==0.63.3
django-celery-beat==2.6.0
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==6.1.0
dnspython==2.6.1
eventlet==0.36.1
flatbuffers==24.3.25
flower==2.0.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
humanize==4.9.0
idna==3.7
kombu==5.3.7
nominatim==0.1
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.7
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.6
requests==2.32.3
requests-cache==1.2.0
retry-requests==2.0.0
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
